Added histogram matching
This module is used to match the histogram of the content image features to the histogram of the style image features. Currently does not support GPU acceleration, however it can help achieve more detailed style images when used in conjunction with gram based style loss, and can style images with minimal impact to content features when used as the sole stylization method.